one action for namespace
